C o m p o n e n t s
FlairComponent
AnnouncerComponent
TownyComponent
Some fixes
Removed old fields and JarUtils
#85
#84
#4
