--- src/main/java/buttondevteam/chat/FlairStates.java
@@ -1,5 +1,0 @@
-package buttondevteam.chat;
-
-public enum FlairStates {
-	Accepted, Ignored, Recognised, Commented, NoComment
-}
