+++ src/main/java/buttondevteam/chat/components/announce/AnnounceCommandBase.java
@@ -1,0 +1,13 @@
+package buttondevteam.chat.components.announce;
+
+import buttondevteam.chat.commands.ucmds.UCommandBase;
+import buttondevteam.lib.chat.CommandClass;
+import buttondevteam.lib.chat.OptionallyPlayerCommandClass;
+
+@CommandClass(modOnly = true)
+@OptionallyPlayerCommandClass(playerOnly = false)
+public abstract class AnnounceCommandBase extends UCommandBase {
+
+	public abstract String[] GetHelpText(String alias);
+
+}
