+++ src/main/java/buttondevteam/chat/components/flair/FlairStates.java
@@ -1,0 +1,5 @@
+package buttondevteam.chat.components.flair;
+
+public enum FlairStates {
+	Accepted, Ignored, Recognised, Commented, NoComment
+}
